+++ core/src/main/java/org/seedstack/seed/core/internal/init/ConsoleManager.java
@@ -12,10 +12,14 @@ import java.io.PrintStream;
 import org.fusesource.jansi.AnsiConsole;
 import org.fusesource.jansi.AnsiPrintStream;
 import org.fusesource.jansi.FilterPrintStream;
+import org.fusesource.jansi.WindowsAnsiPrintStream;
 import org.fusesource.jansi.internal.CLibrary;
 import org.seedstack.seed.ApplicationConfig;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ConsoleManager {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ConsoleManager.class);
     private final PrintStream savedOut = System.out;
     private final PrintStream savedErr = System.err;
 
@@ -28,8 +32,8 @@ public class ConsoleManager {
     }
 
     public synchronized void install(ApplicationConfig.ColorOutput colorOutput) {
-        System.setOut(wrapPrintStream(System.out, CLibrary.STDOUT_FILENO, colorOutput));
-        System.setErr(wrapPrintStream(System.err, CLibrary.STDERR_FILENO, colorOutput));
+        System.setOut(wrapPrintStream(System.out, FileNo.STDOUT, colorOutput, true));
+        System.setErr(wrapPrintStream(System.err, FileNo.STDERR, colorOutput, false));
     }
 
     public synchronized void uninstall() {
@@ -37,32 +41,74 @@ public class ConsoleManager {
         System.setErr(savedErr);
     }
 
-    private PrintStream wrapPrintStream(final PrintStream inputStream, int fileno,
-            ApplicationConfig.ColorOutput colorOutput) {
+    private PrintStream wrapPrintStream(final PrintStream inputStream, FileNo fileno,
+            ApplicationConfig.ColorOutput colorOutput, boolean log) {
         try {
-            if (colorOutput == ApplicationConfig.ColorOutput.PASSTHROUGH || Boolean.getBoolean("jansi.passthrough")) {
+            if (colorOutput == ApplicationConfig.ColorOutput.PASSTHROUGH) {
+                if (log) {
+                    LOGGER.info("Color output passthrough: leaving stdout and stderr untouched");
+                }
                 return inputStream;
-            } else if (colorOutput == ApplicationConfig.ColorOutput.DISABLE || Boolean.getBoolean("jansi.strip")) {
+            } else if (colorOutput == ApplicationConfig.ColorOutput.DISABLE) {
+                if (log) {
+                    LOGGER.info("Color output disabled: stripping stdout and stderr of ANSI codes");
+                }
                 return strippedOutput(inputStream);
             } else if (colorOutput == ApplicationConfig.ColorOutput.ENABLE) {
+                if (log) {
+                    LOGGER.info("Color output enabled: allowing ANSI codes on stdout and stderr");
+                }
                 return ansiOutput(inputStream);
             } else if (colorOutput == ApplicationConfig.ColorOutput.AUTODETECT) {
                 if (isXtermColor()) {
                     // enable color in recognized XTERM color modes
+                    if (log) {
+                        LOGGER.info("XTERM detected: allowing ANSI codes on stdout and stderr");
+                    }
                     return ansiOutput(inputStream);
                 } else if (isIntelliJ()) {
                     // enable color under Intellij
+                    if (log) {
+                        LOGGER.info("IntelliJ detected: allowing ANSI codes on stdout and stderr");
+                    }
                     return ansiOutput(inputStream);
                 } else {
                     // let Jansi handle other detection
-                    return AnsiConsole.wrapPrintStream(inputStream, fileno);
+                    PrintStream result;
+                    switch (fileno) {
+                        case STDOUT:
+                            result = AnsiConsole.wrapPrintStream(inputStream, CLibrary.STDOUT_FILENO);
+                            break;
+                        case STDERR:
+                            result = AnsiConsole.wrapPrintStream(inputStream, CLibrary.STDERR_FILENO);
+                            break;
+                        default:
+                            result = strippedOutput(inputStream);
+                            break;
+                    }
+                    if (log) {
+                        if (result instanceof WindowsAnsiPrintStream) {
+                            LOGGER.info("Windows console detected: using native coloring for stdout and stderr");
+                        } else if (result instanceof AnsiPrintStream) {
+                            LOGGER.info("ANSI not supported: stripping stdout and stderr of ANSI codes");
+                        } else {
+                            LOGGER.info("ANSI supported: allowing ANSI codes on stdout and stderr");
+                        }
+                    }
+                    return result;
                 }
             } else {
                 // Fallback to stripping ANSI codes
+                if (log) {
+                    LOGGER.info("Fallback to stripping ANSI codes");
+                }
                 return strippedOutput(inputStream);
             }
-        } catch (Throwable e) {
+        } catch (Throwable e1) {
             // If any error occurs, strip ANSI codes
+            if (log) {
+                LOGGER.info("Error loading Jansi (unsupported platform ?): stripping stdout and stderr of ANSI codes");
+            }
             return strippedOutput(inputStream);
         }
     }
@@ -110,5 +156,10 @@ public class ConsoleManager {
             ps.flush();
             super.close();
         }
+    }
+
+    private enum FileNo {
+        STDOUT,
+        STDERR
     }
 }